+++ src/main/java/blackjack/domain/card/CardFactory.java
@@ -1,8 +1,5 @@
 package blackjack.domain.card;
 
-import blackjack.domain.card.Card;
-import blackjack.domain.Rank;
-import blackjack.domain.Suit;
 import java.util.ArrayList;
 import java.util.List;
