+++ src/test/java/blackjack/domain/HandTest.java
@@ -3,8 +3,8 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.card.CardFactory;
-import java.util.List;
+import blackjack.domain.card.Rank;
+import blackjack.domain.card.Suit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
